rename cfctx_base in cfctx_basic since generate_makefile.py matches `base`
