+++ .github/workflows/fuzzers.yml
@@ -55,7 +55,7 @@ jobs:
           - aflprefix
           - dropfuzzer
           - aflplusplus_classic_ctx
-          - cfctx_base
+          - cfctx_basic
           - cfctx_bottom
           - cfctx_dataflow_seadsa
           - cfctx_dataflow_svf
