+++ fuzzers/cfctx_dataflow_seadsa/fuzzer.py
@@ -15,7 +15,7 @@
 
 import os
 
-from fuzzers.cfctx_base import fuzzer as aflplusplus_fuzzer
+from fuzzers.cfctx_basic import fuzzer as aflplusplus_fuzzer
 
 
 def build():  # pylint: disable=too-many-branches,too-many-statements
